requirements.txt
crewai
crewai-tools
langchain
duckduckgo-search
python-dotenv
langchain-community
litellm